+++ Sources/CoreCollector/analytics/BitmovinAnalyticsConfig.swift
@@ -218,6 +218,15 @@ public class BitmovinAnalyticsConfig: NSObject {
      */
     public var randomizeUserId = false
 
+    /**
+     * When set to `true`, analytics events that have failed to be sent are cached in a persistent way. Cached elements
+     * from within 14 days are retried at a later point in time, once a network connection is established again. There
+     * will be at most 5,000 elements cached in total and at most 500 elements per playback session.
+     *
+     * Disabling this config flag does not disable the retrying of already cached elements.
+     */
+    public var longTermRetryEnabled = false
+
     public init(key: String, playerKey: String) {
         self.key = key
         self.playerKey = playerKey