+++ Sources/CoreCollector/backend/offline/KeyDerivable.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+internal protocol KeyDerivable {
+    var derivedKey: LosslessStringConvertible? { get }
+}