+++ Sources/CoreCollector/backend/offline/PersistentQueueActor.swift
@@ -1,0 +1,3 @@
+@globalActor internal actor PersistentQueueActor {
+    static let shared = PersistentQueueActor()
+}