+++ Tests/CoreCollectorTests/backend/AuthenticatedDispatcherTests.swift
@@ -6,7 +6,7 @@ import Quick
 import Foundation
 
 class AuthenticatedDispatcherTests: QuickSpec {
-    override func spec() {
+    override class func spec() {
         var mockAuthService = MockAuthenticationService()
         var mockInnerDispatcher = MockEventDataDispatcher()
         var mockNotificationCenter = NotificationCenter()
@@ -80,18 +80,6 @@ class AuthenticatedDispatcherTests: QuickSpec {
             }
         }
         describe("disable") {
-            it("should call innerDispatcher") {
-                // arrange
-                stub(mockInnerDispatcher) { stub in
-                    when(stub.disable()).thenDoNothing()
-                }
-
-                // act
-                dispatcher.disable()
-
-                // assert
-                verify(mockInnerDispatcher).disable()
-            }
             it("should not call add and addAd on innerDispatcher when .disable was called") {
                 // arrange
                 stub(mockInnerDispatcher) { stub in
@@ -108,7 +96,6 @@ class AuthenticatedDispatcherTests: QuickSpec {
                 dispatcher.disable()
 
                 // assert
-                verify(mockInnerDispatcher).disable()
                 dispatcher.addAd(adEventData)
                 dispatcher.add(eventData)
                 verifyNoMoreInteractions(mockInnerDispatcher)
@@ -128,19 +115,29 @@ class AuthenticatedDispatcherTests: QuickSpec {
                 verify(mockInnerDispatcher).resetSourceState()
             }
         }
-        describe("authentication failed") {
-            it("should call innerDispatcher") {
+        describe("authentication denied") {
+            it("should not call add and addAd on innerDispatcher when authentication was denied") {
                 // arrange
                 stub(mockInnerDispatcher) { stub in
                     when(stub.disable()).thenDoNothing()
                 }
 
+                let eventData = EventData("test-impression")
+                let adEventData = AdEventData()
+
+                // emit authentication success event
+                mockNotificationCenter.post(name: .authenticationSuccess, object: mockAuthService)
+
                 // act
-                mockNotificationCenter.post(name: .authenticationFailed, object: mockAuthService)
+                mockNotificationCenter.post(name: .authenticationDenied, object: mockAuthService)
 
                 // assert
-                verify(mockInnerDispatcher).disable()
+                dispatcher.addAd(adEventData)
+                dispatcher.add(eventData)
+                verifyNoMoreInteractions(mockInnerDispatcher)
             }
+        }
+        describe("authentication error") {
             it("should not call add and addAd on innerDispatcher when authentication failed") {
                 // arrange
                 stub(mockInnerDispatcher) { stub in
@@ -154,10 +151,9 @@ class AuthenticatedDispatcherTests: QuickSpec {
                 mockNotificationCenter.post(name: .authenticationSuccess, object: mockAuthService)
 
                 // act
-                mockNotificationCenter.post(name: .authenticationFailed, object: mockAuthService)
+                mockNotificationCenter.post(name: .authenticationError, object: mockAuthService)
 
                 // assert
-                verify(mockInnerDispatcher).disable()
                 dispatcher.addAd(adEventData)
                 dispatcher.add(eventData)
                 verifyNoMoreInteractions(mockInnerDispatcher)