+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development V2
 
+### Added
+- `BitmovinAnalyticsConfig.longTermRetryEnabled` with default value of `false`. When set to `true`, analytics events that have failed to be sent are cached in a persistent way and re-sent as soon as there is network again
+
 ### Fixed 
 - AmazonIVSCollector: `isLive` tracking didn't work